[AIRFLOW-7061] Rename openfass to openfaas (#7721)
