+++ airflow/contrib/hooks/openfaas_hook.py
@@ -15,14 +15,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""This module is deprecated. Please use `airflow.providers.openfass.hooks.openfaas`."""
+"""This module is deprecated. Please use `airflow.providers.openfaas.hooks.openfaas`."""
 
 import warnings
 
 # pylint: disable=unused-import
-from airflow.providers.openfass.hooks.openfaas import OK_STATUS_CODE, OpenFaasHook, requests  # noqa
+from airflow.providers.openfaas.hooks.openfaas import OK_STATUS_CODE, OpenFaasHook, requests  # noqa
 
 warnings.warn(
-    "This module is deprecated. Please use `airflow.providers.openfass.hooks.openfaas`.",
+    "This module is deprecated. Please use `airflow.providers.openfaas.hooks.openfaas`.",
     DeprecationWarning, stacklevel=2
 )
