+++ docs/operators-and-hooks-ref.rst
@@ -1183,7 +1183,7 @@ These integrations allow you to perform various operations using various softwar
 
    * - `OpenFaaS <https://www.openfaas.com/>`__
      -
-     - :mod:`airflow.providers.openfass.hooks.openfaas`
+     - :mod:`airflow.providers.openfaas.hooks.openfaas`
      -
      -
