+++ tests/providers/openfaas/hooks/test_openfaas.py
@@ -25,7 +25,7 @@ import requests_mock
 from airflow.exceptions import AirflowException
 from airflow.hooks.base_hook import BaseHook
 from airflow.models import Connection
-from airflow.providers.openfass.hooks.openfaas import OpenFaasHook
+from airflow.providers.openfaas.hooks.openfaas import OpenFaasHook
 
 FUNCTION_NAME = "function_name"
