+++ tests/test_core_to_contrib.py
@@ -262,7 +262,7 @@ HOOK = [
         'airflow.contrib.hooks.mongo_hook.MongoHook',
     ),
     (
-        'airflow.providers.openfass.hooks.openfaas.OpenFaasHook',
+        'airflow.providers.openfaas.hooks.openfaas.OpenFaasHook',
         'airflow.contrib.hooks.openfaas_hook.OpenFaasHook',
     ),
     (
